refactor(renderer): decompose App.tsx into feature slices + hooks
Quality-review slice 3 (behavior-preserving). App.tsx 1423 -> 746 lines:
- auth/SignInPanel.tsx: the sign-in feature UI (SignInPanel + card +
  loading row) moves into the auth slice that already owned its reducer.
- settings/: new slice — SettingsView + AccountSettings + Environment.
- shell/EmptyState.tsx + shell/Outlet.tsx (TransientOutlet, ColdOutlet):
  the outlet render helpers become presentational components.
- connection/use-thread-controls.ts: the optimistic-apply/revert/cache
  controls choreography (#66/#70/#72/#75) as one hook — the shared revert
  loop exists once and the stale-closure mirror ref no longer leaks into App.
- connection/use-workspace-actions.ts: delete/remove/flags/rename multi-
  store reconciliation behind a seam drivable with fake dispatchers.
- connection/resolve-controls.ts: the picker's four-way fallback chain
  (and its localStorage read) extracted from render JSX into a pure fn.
