+++ src/renderer/src/App.tsx
@@ -3,19 +3,11 @@ import type {
   AuthMethod,
   ListMetadataResult,
   StartThreadResult,
-  ThreadAgentControls,
-  ThreadConfigAxis,
   ThreadConnection,
   ThreadMeta,
   VibeDetectResult,
 } from '../../shared/ipc'
-import {
-  authReducer,
-  initialAuthViewState,
-  selectAuthView,
-  signedInAuthViewState,
-} from './auth/auth-view'
-import { routeThreadResult, type ConnectState } from './connection/routing'
+import { routeThreadResult } from './connection/routing'
 import {
   agentIdOf,
   connectedWorkspaceIds,
@@ -25,51 +17,32 @@ import {
   shouldConnect,
 } from './connection/connections'
 import {
-  boundConfigValue,
-  configFor,
-  currentConfigValue,
-  draftControls,
   initialWorkspaceThreads,
-  reassertions,
-  selectedFor,
   workspaceThreadsReducer,
   workspaceThreadStateFor,
   type WorkspaceThreadState,
 } from './connection/workspace-threads'
 import { ConnectedWorkspace } from './connection/ConnectedWorkspace'
 import { routeThreadSelection, seedSessionId } from './connection/thread-selection'
-import { ColdThread } from './conversation/ColdThread'
-import {
-  clearThreadStatus,
-  setThreadStatus,
-  type ThreadStatusMap,
-} from './conversation/thread-status'
-import { clearDraft } from './conversation/composer-draft-store'
-import {
-  getWorkspaceControls,
-  setWorkspaceControls,
-  workspaceControlsKey,
-} from './connection/workspace-controls-store'
+import { useThreadControls } from './connection/use-thread-controls'
+import { useWorkspaceActions } from './connection/use-workspace-actions'
+import { resolveActiveControls } from './connection/resolve-controls'
+import { setThreadStatus, type ThreadStatusMap } from './conversation/thread-status'
+import { setWorkspaceControls, workspaceControlsKey } from './connection/workspace-controls-store'
 import { ArrowLeft, ArrowRight, Maximize2, PanelLeft, PanelRight, Terminal } from 'lucide-react'
-import { Button } from './ui/button'
 import { IconButton } from './ui/icon-button'
-import { Card } from './ui/card'
 import { Shell, type WorkspaceFlags } from './shell/Shell'
-import { Logo } from './shell/logo'
-import { LogoSnakeSpinner } from './shell/logo-snake-spinner'
-import { heroHeadline } from './shell/hero-headline'
-import { firstRunState, type FirstRunState } from './shell/first-run'
-import { findSelectedThread, initialNavState, navReducer, type NavState } from './shell/nav-reducer'
+import { firstRunState } from './shell/first-run'
+import { initialNavState, navReducer } from './shell/nav-reducer'
 import {
   getSidebarCollapsed,
   setSidebarCollapsed as setSidebarCollapsedStore,
 } from './shell/sidebar-collapsed-store'
-import {
-  removeWorkspacePanel,
-  toggleWorkspacePanelVisibility,
-  useWorkspacePanel,
-} from './side-panel/side-panel-store'
+import { toggleWorkspacePanelVisibility, useWorkspacePanel } from './side-panel/side-panel-store'
 import { deriveUnifiedThreads, workspaceFlags, type UnifiedThreadRow } from './shell/unified-threads'
+import { EmptyState } from './shell/EmptyState'
+import { ColdOutlet, TransientOutlet } from './shell/Outlet'
+import { SettingsView } from './settings/SettingsView'
 
 /** A stable empty live-set for Workspaces with no live-state yet (no re-alloc). */
 const NO_LIVE: ReadonlySet<string> = new Set()
@@ -90,6 +63,10 @@ const NO_LIVE: ReadonlySet<string> = new Set()
  * never-connected Workspace replays correctly even after another connected
  * (TB1-review finding 2); the sidebar suppresses per-Thread highlights for a
  * connected Workspace, whose live view owns Thread selection (finding 1).
+ *
+ * The Workspace/Thread lifecycle mutations live in `useWorkspaceActions` and the
+ * per-Thread agent-controls choreography in `useThreadControls` — App wires live
+ * state into them; the auth/settings/empty-state UIs live in their feature slices.
  */
 export function App(): JSX.Element {
   const [detect, setDetect] = useState<VibeDetectResult | null>(null)
@@ -138,11 +115,10 @@ export function App(): JSX.Element {
   // user has since switched away). Mirrors the latest rendered nav selection.
   const selectionRef = useRef<string | null>(nav.selectedWorkspaceId)
   selectionRef.current = nav.selectedWorkspaceId
-  // Latest workspace-threads, mirrored into a ref so the async `onBound` re-assert
-  // (#72) reads the CURRENT selection cache — its render-time closure is stale by the
-  // time a resume's `thread:bound` fires (mirrors `selectionRef` above).
-  const workspaceThreadsRef = useRef(workspaceThreads)
-  workspaceThreadsRef.current = workspaceThreads
+
+  // The per-Thread agent-controls choreography (#66/#70/#72/#75, ADR-0007): the
+  // optimistic-apply/revert/cache idiom + the stale-closure mirror ref live in the hook.
+  const controls = useThreadControls(workspaceThreads, wtDispatch)
 
   async function runDetect(): Promise<void> {
     setLoading(true)
@@ -156,110 +132,32 @@ export function App(): JSX.Element {
   }
 
   /**
-   * Delete a Thread from the unified list (TB6 + #48 safe-delete). Main removes its
-   * metadata + JSONL and best-effort closes any live session. The sidebar only
-   * offers delete for a row `isThreadDeletable` proves safe (a cold row, or any
-   * idle non-primary live row — its real per-Thread streaming is now observable for
-   * ALL live Threads via main's push, #53), so we never tear a Thread out mid-stream.
-   *
-   * Reselection is gated on SELECTION, not liveness: `active`/`nav.selectedThreadId`
-   * can legitimately point at a COLD (history) row, and dropping it from `recents`
-   * would leave the outlet/sidebar pinned to a now-gone Thread. So whenever the
-   * deleted Thread is the active/selected one of a CONNECTED Workspace, reselect the
-   * connection's (always-live) primary Thread. The `wt remove` (drop from live-state)
-   * runs ONLY when it was live; its stale status entry is cleared either way.
-   *
-   * Main re-validates streaming authoritatively and can REFUSE a delete that raced a
-   * just-started turn (`{ok:false, reason:'streaming'}`, #53); we bail and leave the
-   * row in place so the UI never drops a Thread main still hosts mid-stream.
+   * Select a Thread from the sidebar (TB3 #48): pin it in nav (the single source of
+   * truth) and, for a connected Workspace, remember it as the active (kept-mounted)
+   * Thread so backgrounding the Workspace keeps it streaming.
    */
-  async function deleteThread(thread: ThreadMeta): Promise<void> {
-    const result = await window.api.deleteThread(thread.id)
-    if (!result.ok) return
-    const wts = workspaceThreadStateFor(workspaceThreads, thread.workspaceId)
-    if (wts?.live.has(thread.id)) {
-      wtDispatch({ type: 'remove', workspaceId: thread.workspaceId, threadId: thread.id })
+  function selectThreadInWorkspace(workspaceId: string, threadId: string): void {
+    navDispatch({ type: 'select-thread', workspaceId, threadId })
+    if (workspaceThreadStateFor(workspaceThreads, workspaceId)) {
+      wtDispatch({ type: 'select', workspaceId, threadId })
     }
-    const conn = connections[thread.workspaceId]
-    const wasSelected = wts?.active === thread.id || nav.selectedThreadId === thread.id
-    if (conn?.status === 'connected' && wasSelected) {
-      selectThreadInWorkspace(thread.workspaceId, conn.thread.threadId)
-    }
-    setStatuses((prev) => clearThreadStatus(prev, thread.id))
-    // Drop the deleted Thread's persisted composer draft (#60) — no orphaned text.
-    clearDraft(window.localStorage, thread.id)
-    await refreshRecents()
   }
 
-  /**
-   * Remove a Workspace from the sidebar ("Remove project", Codex-style). Main stops
-   * its warm agent (if any — allowed even mid-turn) and removes OUR records (the
-   * Workspace + Thread metadata + JSONL); it NEVER deletes files on disk. Here we
-   * reconcile local state so the project disappears cleanly:
-   *  - If it was the selected Workspace, clear the nav selection (lands on the empty
-   *    state); leave the selection untouched when removing a non-selected project.
-   *  - Drop its connection and per-Workspace live-state. Both are idempotent: for a
-   *    CONNECTED project main disposed the agent and pushed `agent:evicted`, whose
-   *    handler already dropped the connection by agentId — so `clear` is a no-op then
-   *    (no double-removal), and it covers the cold/unconnected case that evict misses.
-   *  - Drop each removed Thread's persisted composer draft + renderer status, mirroring
-   *    `deleteThread` — so a removed project leaves no orphaned localStorage/status keys.
-   *  - `refreshRecents()` LAST, dropping it from the persisted list the sidebar renders.
-   */
-  async function removeWorkspace(workspaceId: string): Promise<void> {
-    await window.api.removeWorkspace(workspaceId)
-    // Snapshot the removed Workspace's Thread ids from the CURRENT list, before the
-    // refresh drops it, so we can clear their renderer-local residue below.
-    const removedThreadIds = recents.find((w) => w.id === workspaceId)?.threads.map((t) => t.id) ?? []
-    if (nav.selectedWorkspaceId === workspaceId) navDispatch({ type: 'clear' })
-    connDispatch({ type: 'clear', workspaceId })
-    wtDispatch({ type: 'remove-workspace', workspaceId })
-    if (removedThreadIds.length > 0) {
-      setStatuses((prev) => removedThreadIds.reduce((acc, id) => clearThreadStatus(acc, id), prev))
-      for (const id of removedThreadIds) clearDraft(window.localStorage, id)
-    }
-    // Drop the side-panel entry too (#193): workspaceIds are fresh UUIDs, so a removed
-    // Workspace's open-tabs blob would otherwise sit unreachable in localStorage forever.
-    removeWorkspacePanel(workspaceId)
-    await refreshRecents()
-  }
-
-  /**
-   * Toggle a Thread's persisted per-Thread flags (#132 pin / #133 archive). A SAFE
-   * metadata op — no session teardown — so it runs on any row (active or cold-peek).
-   * Best-effort in main (ADR-0005); we refresh the recents list so the new flag
-   * reflects in the sidebar's derivation (`orderByPin` / `partitionArchived`). A
-   * `{ok:false}` (store failure) leaves the list as-is — the toggle is a no-op.
-   */
-  async function setThreadFlags(
-    threadId: string,
-    flags: { pinned?: boolean; archived?: boolean },
-  ): Promise<void> {
-    const result = await window.api.setThreadFlags({ threadId, ...flags })
-    if (!result.ok) return
-    await refreshRecents()
-  }
-
-  /**
-   * Rename a Thread. Main owns the title in OUR store, and additionally syncs the
-   * vibe-acp side when the Thread is live — so we pass the hosting `agentId` (when its
-   * Workspace is connected) and the Thread's bound `sessionId`; main no-ops the ACP
-   * call for a cold Thread. Refresh the cold list on success so the sidebar re-labels
-   * (the setter holds list position, so the Thread doesn't jump). A `{ok:false}`
-   * (empty title / store failure) leaves the label unchanged.
-   */
-  async function renameThread(thread: ThreadMeta, title: string): Promise<void> {
-    const conn = connections[thread.workspaceId]
-    const agentId = conn ? agentIdOf(conn) : null
-    const result = await window.api.setThreadTitle({
-      threadId: thread.id,
-      title,
-      agentId: agentId ?? undefined,
-      sessionId: thread.sessionId,
-    })
-    if (!result.ok) return
-    await refreshRecents()
-  }
+  // The Workspace/Thread lifecycle mutations (delete / remove-project / flags /
+  // rename): multi-store reconciliation behind one seam (use-workspace-actions.ts).
+  const actions = useWorkspaceActions({
+    recents,
+    nav,
+    connections,
+    workspaceThreads,
+    navDispatch,
+    connDispatch,
+    wtDispatch,
+    setStatuses,
+    refreshRecents,
+    selectThreadInWorkspace,
+    storage: window.localStorage,
+  })
 
   /**
    * Record a Workspace connect outcome: set its ConnectState and, when connected,
@@ -269,7 +167,8 @@ export function App(): JSX.Element {
    * the outlet routes to it, without yanking focus from a Workspace switched-to
    * while this connect was in flight.
    */
-  function applyConnectResult(workspaceId: string, state: ConnectState, focus: boolean): void {
+  function applyConnectResult(workspaceId: string, result: StartThreadResult, focus: boolean): void {
+    const state = routeThreadResult(result)
     connDispatch({ type: 'set', workspaceId, state })
     if (state.status !== 'connected') return
     // Seed the connect-time (primary) Thread's controls per-Thread (#70) from the
@@ -291,100 +190,6 @@ export function App(): JSX.Element {
     if (focus || selectionRef.current === workspaceId) {
       navDispatch({ type: 'select-thread', workspaceId, threadId: state.thread.threadId })
     }
-  }
-
-  /**
-   * Select a Thread from the sidebar (TB3 #48): pin it in nav (the single source of
-   * truth) and, for a connected Workspace, remember it as the active (kept-mounted)
-   * Thread so backgrounding the Workspace keeps it streaming.
-   */
-  function selectThreadInWorkspace(workspaceId: string, threadId: string): void {
-    navDispatch({ type: 'select-thread', workspaceId, threadId })
-    if (workspaceThreadStateFor(workspaceThreads, workspaceId)) {
-      wtDispatch({ type: 'select', workspaceId, threadId })
-    }
-  }
-
-  /**
-   * Change an agent control for ONE Thread (#66/#70, ADR-0007): reflect the pick
-   * OPTIMISTICALLY on THAT Thread's per-Thread config (keyed by `threadId` in the
-   * workspace-threads store — a change emits no notification, so the `{}` result is
-   * the only signal), fire the IPC, and on an `{ok:false}` REVERT to the value shown
-   * before — leaving the control displaying the agent's real state — and surface the
-   * error. Reads the prior value from the per-Thread config up front so the revert is
-   * exact; a sibling Thread's controls are never touched.
-   */
-  function changeThreadConfig(
-    workspaceId: string,
-    agentId: string,
-    threadId: string,
-    axis: ThreadConfigAxis,
-    value: string,
-    sessionId: string,
-  ): void {
-    const prev = currentConfigValue(workspaceThreads, workspaceId, threadId, axis)
-    if (prev === value) return // already current — no optimistic churn, no IPC round-trip
-    wtDispatch({ type: 'set-config', workspaceId, threadId, axis, value })
-    void window.api.setThreadConfig({ agentId, sessionId, axis, value }).then((res) => {
-      if (res.ok) {
-        // Remember the CONFIRMED pick (#72) so a later resume (re-warm / cold continue)
-        // re-asserts it — Vibe resets Mode to `default` on `session/load`. Cached only
-        // here, never on the optimistic update or the revert below.
-        wtDispatch({ type: 'cache-selection', workspaceId, threadId, axis, value })
-        return
-      }
-      console.error(`Failed to set ${axis} to "${value}": ${res.error}`)
-      if (prev !== null) wtDispatch({ type: 'set-config', workspaceId, threadId, axis, value: prev })
-    })
-  }
-
-  /**
-   * Re-assert a Thread's cached selection after a resume (#72). Vibe resets Mode to
-   * `default` on `session/load`, so a Thread whose session was lost (idle-evicted +
-   * re-warmed per TB5, or a cold continue) and resumed reports its DEFAULT controls on
-   * `thread:bound`. For each axis whose cached `selected` differs from the resumed
-   * value, optimistically reflect it on the displayed config AND fire the IPC to put
-   * the live session back to the user's choice — reverting (to the resumed value) +
-   * logging on failure, mirroring `changeThreadConfig`. Reads the cache from the ref so
-   * an async resume sees the latest, not its stale render-time closure. No-ops for a
-   * fresh mint (no cache) and when the resumed value already matches.
-   */
-  function reassertAfterResume(
-    workspaceId: string,
-    agentId: string,
-    threadId: string,
-    sessionId: string,
-    controls: ThreadAgentControls,
-  ): void {
-    const selected = selectedFor(workspaceThreadsRef.current, workspaceId, threadId)
-    for (const { axis, value } of reassertions(selected, controls)) {
-      const prev = boundConfigValue(controls, axis) // the resumed value to revert to
-      wtDispatch({ type: 'set-config', workspaceId, threadId, axis, value })
-      void window.api.setThreadConfig({ agentId, sessionId, axis, value }).then((res) => {
-        if (res.ok) return
-        console.error(`Failed to re-assert ${axis} to "${value}": ${res.error}`)
-        if (prev !== null) wtDispatch({ type: 'set-config', workspaceId, threadId, axis, value: prev })
-      })
-    }
-  }
-
-  /**
-   * Pre-select an agent control on a New-thread DRAFT (#75), before its first prompt
-   * binds a session. A draft has no live session, so there's NO IPC — we only cache
-   * the pick into the in-memory `selected` map (keyed by `threadId`). The display
-   * updates because the draft's picker reads `draftControls(connection, selected)`, and
-   * because this writes the SAME cache `changeThreadConfig` writes on a bound Thread,
-   * the EXISTING `reassertAfterResume` (#72) applies the pre-pick to the session the
-   * instant the first prompt mints it — no second apply path, and no residue (the cache
-   * evaporates with the draft / on restart, ADR-0007).
-   */
-  function preselectDraftConfig(
-    workspaceId: string,
-    threadId: string,
-    axis: ThreadConfigAxis,
-    value: string,
-  ): void {
-    wtDispatch({ type: 'cache-selection', workspaceId, threadId, axis, value })
   }
 
   /**
@@ -519,7 +324,7 @@ export function App(): JSX.Element {
     connDispatch({ type: 'set', workspaceId, state: { status: 'connecting', workspaceDir: workspace.dir } })
     try {
       const result = await window.api.startThread({ workspaceDir: workspace.dir })
-      applyConnectResult(workspaceId, routeThreadResult(result), false)
+      applyConnectResult(workspaceId, result, false)
       void refreshRecents()
     } finally {
       connectingRef.current.delete(workspaceId)
@@ -547,7 +352,7 @@ export function App(): JSX.Element {
       setRecents(list)
       const ws = list.find((w) => w.dir === workspaceDir)
       if (!ws) {
-        // Degraded (no store / failed list): we can't key or select this
+        // Degraded (a failed persist / list): we can't key or select this
         // connection, so dispose the just-spawned agent rather than leak a warm
         // connected child the renderer can never reach until quit.
         const agentId = agentIdOfResult(result)
@@ -555,7 +360,7 @@ export function App(): JSX.Element {
         return
       }
       navDispatch({ type: 'select-workspace', workspaceId: ws.id })
-      applyConnectResult(ws.id, routeThreadResult(result), true)
+      applyConnectResult(ws.id, result, true)
     } finally {
       setOpening(false)
     }
@@ -564,7 +369,7 @@ export function App(): JSX.Element {
   // After sign-in (or in-place re-auth) the Workspace's warm agent is already
   // started + signed in; open a Thread on it and land in a connected conversation.
   async function continueToThread(workspaceId: string, agentId: string): Promise<void> {
-    applyConnectResult(workspaceId, routeThreadResult(await window.api.openThread({ agentId })), true)
+    applyConnectResult(workspaceId, await window.api.openThread({ agentId }), true)
     void refreshRecents()
   }
 
@@ -583,7 +388,7 @@ export function App(): JSX.Element {
     const result = await window.api.startThread({ workspaceDir: workspace.dir, continueThreadId: thread.id })
     // Main opened NO extra Thread — the continued Thread IS the connection Thread, so
     // `applyConnectResult` seeds it live + selects it (its first prompt drives resume).
-    applyConnectResult(workspace.id, routeThreadResult(result), true)
+    applyConnectResult(workspace.id, result, true)
     void refreshRecents()
   }
 
@@ -664,66 +469,46 @@ export function App(): JSX.Element {
       // A controlled outlet now (TB3 #48): the sidebar drives selection; this just
       // renders App's chosen active Thread live or cold.
       <ConnectedWorkspace
-          key={conn.agentId}
-          connection={conn}
-          activeThread={activeThread}
-          isLive={isLive}
-          isActive={isActive}
-          busy={busy}
-          seedSessionId={seed}
-          controls={
-            // A bound Thread sources its OWN live config (#70); a draft (no config
-            // seeded) shows the connection's option lists + defaults, overlaid with any
-            // cached pre-pick (#75). CAVEAT: a CONTINUED (reopened, not-yet-bound) Thread
-            // also has no config, so it shows the connection DEFAULTS too — honest for
-            // Mode (session/load resets it to default) but the MODEL can persist across
-            // a resume (acp-capture §10), so a reopened Thread may briefly show the
-            // default model until its first prompt's bind reports the real one and
-            // self-corrects. We don't eagerly resume to learn it (#33 defers load to the
-            // first prompt); Model isn't trust-relevant (it doesn't gate writes), so the
-            // transient pre-prompt mismatch is accepted.
-            // A never-bound draft's connection advertises all-null controls (ADR-0011
-            // opens no session until the first prompt), so before falling back to the
-            // connection we try the per-Workspace cache (#153) — the last bound session's
-            // option lists — so the picker shows immediately instead of after send.
-            configFor(workspaceThreads, conn.workspaceId, activeThread.id) ??
-            draftControls(
-              getWorkspaceControls(
-                window.localStorage,
-                workspaceControlsKey(conn.workspaceId, conn.workspaceDir),
-              ) ?? connectionControlsOf(conn),
-              selectedFor(workspaceThreads, conn.workspaceId, activeThread.id),
+        key={conn.agentId}
+        connection={conn}
+        activeThread={activeThread}
+        isLive={isLive}
+        isActive={isActive}
+        busy={busy}
+        seedSessionId={seed}
+        controls={resolveActiveControls(workspaceThreads, conn, activeThread.id, window.localStorage)}
+        onSetConfig={(axis, value, sessionId) => {
+          // A real session => the bound IPC path (#70); a null session => a draft
+          // pre-pick that only caches (#75), applied to the session on first bind.
+          if (sessionId) {
+            controls.changeThreadConfig(conn.workspaceId, conn.agentId, activeThread.id, axis, value, sessionId)
+          } else {
+            controls.preselectDraftConfig(conn.workspaceId, activeThread.id, axis, value)
+          }
+        }}
+        onBound={(sessionId, boundControls) => {
+          // Seed the displayed config from the bound session's reported values, then
+          // re-assert the user's cached selection over them (#72) — a resume reports
+          // defaults, so this restores a prior non-default Mode/Model/effort.
+          wtDispatch({ type: 'bind', workspaceId: conn.workspaceId, threadId: activeThread.id, sessionId, controls: boundControls })
+          if (boundControls) {
+            controls.reassertAfterResume(conn.workspaceId, conn.agentId, activeThread.id, sessionId, boundControls)
+            // Cache the bound session's option lists per Workspace (#153) so the NEXT
+            // never-bound draft here shows the picker before its own first prompt binds.
+            setWorkspaceControls(
+              window.localStorage,
+              workspaceControlsKey(conn.workspaceId, conn.workspaceDir),
+              boundControls,
             )
           }
-          onSetConfig={(axis, value, sessionId) => {
-            // A real session => the bound IPC path (#70); a null session => a draft
-            // pre-pick that only caches (#75), applied to the session on first bind.
-            if (sessionId) changeThreadConfig(conn.workspaceId, conn.agentId, activeThread.id, axis, value, sessionId)
-            else preselectDraftConfig(conn.workspaceId, activeThread.id, axis, value)
-          }}
-          onBound={(sessionId, controls) => {
-            // Seed the displayed config from the bound session's reported values, then
-            // re-assert the user's cached selection over them (#72) — a resume reports
-            // defaults, so this restores a prior non-default Mode/Model/effort.
-            wtDispatch({ type: 'bind', workspaceId: conn.workspaceId, threadId: activeThread.id, sessionId, controls })
-            if (controls) {
-              reassertAfterResume(conn.workspaceId, conn.agentId, activeThread.id, sessionId, controls)
-              // Cache the bound session's option lists per Workspace (#153) so the NEXT
-              // never-bound draft here shows the picker before its own first prompt binds.
-              setWorkspaceControls(
-                window.localStorage,
-                workspaceControlsKey(conn.workspaceId, conn.workspaceDir),
-                controls,
-              )
-            }
-          }}
-          onContinue={() => {
-            wtDispatch({ type: 'open', workspaceId: conn.workspaceId, threadId: activeThread.id })
-            navDispatch({ type: 'select-thread', workspaceId: conn.workspaceId, threadId: activeThread.id })
-          }}
-          onCloseCold={() => selectThreadInWorkspace(conn.workspaceId, conn.threadId)}
-          onAuthExpired={(authMethods) => toSignInPanel(conn.workspaceId, conn.agentId, conn.workspaceDir, authMethods)}
-        />
+        }}
+        onContinue={() => {
+          wtDispatch({ type: 'open', workspaceId: conn.workspaceId, threadId: activeThread.id })
+          navDispatch({ type: 'select-thread', workspaceId: conn.workspaceId, threadId: activeThread.id })
+        }}
+        onCloseCold={() => selectThreadInWorkspace(conn.workspaceId, conn.threadId)}
+        onAuthExpired={(authMethods) => toSignInPanel(conn.workspaceId, conn.agentId, conn.workspaceDir, authMethods)}
+      />
     )
   }
 
@@ -738,10 +523,9 @@ export function App(): JSX.Element {
   // permission surfaces in the sidebar (and its delete gate) without a mounted view.
   // The on-demand Settings page (#130): a top-level outlet view, routed by the nav
   // reducer (`nav.view === 'settings'`), that hosts the env/CLI detection status the
-  // sidebar gear used to toggle (reusing the same `Environment` component + `detect`).
-  // It swaps the conversation/cold/empty outlet WITHOUT unmounting the connected
-  // Workspaces — they stay mounted-but-hidden so a background turn keeps streaming, and
-  // closing Settings (or picking a project/thread) returns to exactly the same view.
+  // sidebar gear used to toggle. It swaps the conversation/cold/empty outlet WITHOUT
+  // unmounting the connected Workspaces — they stay mounted-but-hidden so a background
+  // turn keeps streaming, and closing Settings returns to exactly the same view.
   const inSettings = nav.view === 'settings'
   const outlet = (
     <>
@@ -786,18 +570,19 @@ export function App(): JSX.Element {
           }}
         />
       ) : selected.status === 'connected' ? null : ( // connected: rendered (visible) in the keep-mounted map above
-        selected.status !== 'idle'
-          ? renderTransientOutlet(selected, {
-              continueToThread: (agentId) => void continueToThread(selectedWs ?? '', agentId),
-              onRetry: () => selectedWs && void connectWorkspace(selectedWs),
-            })
-          : renderColdOutlet(
-              recents,
-              nav,
-              {
-                onClose: () => navDispatch({ type: 'clear' }),
-                onContinue: (thread) => void continueColdThread(thread),
-              },
+        selected.status !== 'idle' ? (
+          <TransientOutlet
+            connect={selected}
+            onContinueToThread={(agentId) => void continueToThread(selectedWs ?? '', agentId)}
+            onRetry={() => selectedWs && void connectWorkspace(selectedWs)}
+          />
+        ) : (
+          <ColdOutlet
+            recents={recents}
+            nav={nav}
+            onClose={() => navDispatch({ type: 'clear' })}
+            onContinue={(thread) => void continueColdThread(thread)}
+            empty={
               <EmptyState
                 state={firstRunState(detect, recents)}
                 detect={detect}
@@ -806,8 +591,10 @@ export function App(): JSX.Element {
                 workspaceName={selectedWorkspaceName}
                 onRecheck={() => void runDetect()}
                 onOpenProject={() => void openProject()}
-              />,
-            )
+              />
+            }
+          />
+        )
       )}
     </>
   )
@@ -887,10 +674,10 @@ export function App(): JSX.Element {
         onSelectThread={selectThreadInWorkspace}
         onNewThread={startNewChat}
         onNewThreadInWorkspace={newThreadInWorkspace}
-        onDeleteThread={deleteThread}
-        onRemoveWorkspace={removeWorkspace}
-        onSetThreadFlags={setThreadFlags}
-        onRenameThread={renameThread}
+        onDeleteThread={actions.deleteThread}
+        onRemoveWorkspace={actions.removeWorkspace}
+        onSetThreadFlags={actions.setThreadFlags}
+        onRenameThread={actions.renameThread}
         onOpenSettings={() => navDispatch({ type: 'open-settings' })}
       />
     </div>
@@ -942,20 +729,10 @@ function synthConnectionMeta(conn: ThreadConnection): ThreadMeta {
 }
 
 /**
- * Project a connection's advertised agent-controls (#75): the connect-time option
- * lists + DEFAULT current values, never optimistically mutated (a pick lands in
- * `workspace-threads.config`, not here — #70). Used to seed a draft's picker so it
- * shows the agent defaults a default-mint would produce.
- */
-function connectionControlsOf(conn: ThreadConnection): ThreadAgentControls {
-  return { modes: conn.modes, models: conn.models, reasoningEffort: conn.reasoningEffort }
-}
-
-/**
  * The pool-minted `agentId` a `startThread` result carries, when any — present on
  * a connected (`thread.agentId`) or not-signed-in (`agentId`) result, absent on a
  * non-auth error (main already disposed that agent). Used to dispose a warm agent
- * the renderer can't key in degraded mode (no store).
+ * the renderer can't key when the Workspace record couldn't be resolved.
  */
 function agentIdOfResult(result: StartThreadResult): string | null {
   if (result.ok) return result.thread.agentId
@@ -963,461 +740,7 @@ function agentIdOfResult(result: StartThreadResult): string | null {
   return null
 }
 
-/**
- * The selected Workspace's NON-connected outlet state (connecting / not-signed-in /
- * error). Only the selected Workspace shows a transient view; connected Workspaces
- * render via the keep-mounted map instead.
- */
-function renderTransientOutlet(
-  connect: ConnectState,
-  handlers: { continueToThread: (agentId: string) => void; onRetry: () => void },
-): ReactNode {
-  switch (connect.status) {
-    case 'connecting':
-      return (
-        <div className="mx-auto mt-14 flex max-w-[420px] flex-col items-center gap-2 text-center">
-          <span className="dot dot--pending" aria-hidden />
-          <div className="text-sm font-semibold text-text-strong">Connecting…</div>
-          <div className="text-[13px] leading-relaxed text-muted">
-            Launching <code>vibe-acp</code> in <code>{connect.workspaceDir}</code> and running the
-            ACP handshake.
-          </div>
-        </div>
-      )
-    case 'not-signed-in':
-      return (
-        <SignInPanel
-          key={connect.agentId}
-          agentId={connect.agentId}
-          authMethods={connect.authMethods}
-          onSignedIn={() => handlers.continueToThread(connect.agentId)}
-        />
-      )
-    case 'error':
-      return (
-        <div className="alert">
-          <div className="alert__title">Couldn’t connect</div>
-          <div className="alert__message">{connect.message}</div>
-          {connect.hint && <div className="alert__hint">{connect.hint}</div>}
-          <button className="btn alert__action" onClick={handlers.onRetry}>
-            Retry
-          </button>
-        </div>
-      )
-    default:
-      return null
-  }
-}
-
-/**
- * The idle (no live agent) outlet for the selected Workspace: the nav-selected cold
- * Thread replayed read-only from JSONL (TB3) with a Continue affordance (TB4), or a
- * placeholder when nothing is selected. Reached only when the selected Workspace has
- * no connection — so a cold click after another Workspace connected still routes here.
- */
-function renderColdOutlet(
-  recents: ListMetadataResult,
-  nav: NavState,
-  handlers: { onClose: () => void; onContinue: (thread: ThreadMeta) => void },
-  empty: ReactNode,
-): ReactNode {
-  const selectedThread = findSelectedThread(recents, nav)
-  if (!selectedThread) return empty
-  return (
-    <ColdThread
-      key={selectedThread.id}
-      thread={selectedThread}
-      onClose={handlers.onClose}
-      onContinue={() => handlers.onContinue(selectedThread)}
-    />
-  )
-}
-
-/**
- * The first-run / empty outlet shown when nothing is connected or selected (#49).
- * Driven by the pure `firstRunState`: when `vibe` / `vibe-acp` is missing the env
- * status is surfaced PROMINENTLY here (the user can't proceed until it's installed);
- * when the toolchain's present but no Workspaces exist it nudges Open-project; once
- * everything's set up it's a neutral placeholder (env tucked behind settings).
- */
-function EmptyState({
-  state,
-  detect,
-  loading,
-  opening,
-  workspaceName,
-  onRecheck,
-  onOpenProject,
-}: {
-  state: FirstRunState
-  detect: VibeDetectResult | null
-  loading: boolean
-  opening: boolean
-  /** The selected Workspace's name, emphasized in the idle hero headline (or null). */
-  workspaceName: string | null
-  onRecheck: () => void
-  onOpenProject: () => void
-}): JSX.Element {
-  if (state === 'needs-install') {
-    return (
-      <div className="flex max-w-[460px] flex-col items-start gap-3">
-        <div className="text-[15px] font-semibold text-text-strong">
-          Install Mistral Vibe to get started
-        </div>
-        <p className="hint">
-          vibe-mistro drives the <code>vibe-acp</code> ACP server. Install the Mistral Vibe CLI and{' '}
-          <code>vibe-acp</code>, then re-check below.
-        </p>
-        <Environment detect={detect} loading={loading} onRecheck={onRecheck} />
-      </div>
-    )
-  }
-  if (state === 'no-workspaces') {
-    return (
-      <div className="flex max-w-[460px] flex-col items-start gap-3">
-        <div className="text-[15px] font-semibold text-text-strong">No workspaces yet</div>
-        <p className="hint">Open a project to spawn its agent and start a thread.</p>
-        <Button onClick={onOpenProject} disabled={opening}>
-          {opening ? 'Connecting…' : 'Open project'}
-        </Button>
-      </div>
-    )
-  }
-  // idle — the empty-state hero: a centered logo + a dynamic headline with the
-  // selected Workspace name in orange (`--accent-emphasis`).
-  const headline = heroHeadline(workspaceName)
-  return (
-    <div className="mx-auto flex h-full max-w-[830px] flex-col items-center justify-center gap-6 text-center">
-      <Logo size={52} />
-      <h1 className="text-[37px] font-semibold tracking-[-0.6px] text-text-strong">
-        {headline.lead}
-        {headline.name && <span className="text-accent-emphasis">{headline.name}</span>}
-        {headline.tail}
-      </h1>
-      <p className="hint">
-        Select a thread from the sidebar to view it, or open a project to start a live agent.
-      </p>
-    </div>
-  )
-}
-
-/**
- * The Settings page (#130): an on-demand, nav-routed outlet view that replaces the
- * old sidebar gear. A titled panel with a back/close affordance (`onClose` dispatches
- * `close-settings`, so you can leave even with nothing selected) hosting the existing
- * `Environment` env/CLI status. Future settings land here. This is an ADDITIONAL place
- * to check the toolchain — NOT a replacement for the first-run `EmptyState`, which still
- * surfaces a missing toolchain prominently in the outlet when nothing's installed.
- */
-function SettingsView({
-  detect,
-  loading,
-  onRecheck,
-  onClose,
-  account,
-  onSignedOut,
-}: {
-  detect: VibeDetectResult | null
-  loading: boolean
-  onRecheck: () => void
-  onClose: () => void
-  /** The selected Workspace's signed-in account, or null when none is connected. */
-  account: AccountInfo | null
-  onSignedOut: (authMethods: AuthMethod[]) => void
-}): JSX.Element {
-  return (
-    <div className="mx-auto flex w-full max-w-[560px] flex-col gap-5">
-      <div className="flex items-center gap-2">
-        <IconButton aria-label="Back" title="Back" onClick={onClose}>
-          <ArrowLeft className="size-4" aria-hidden />
-        </IconButton>
-        <h1 className="text-[19px] font-semibold tracking-tight text-text-strong">Settings</h1>
-      </div>
-      <section className="flex flex-col gap-2">
-        <h2 className="text-[13px] font-semibold text-faint">Account</h2>
-        <AccountSettings
-          // Key by agentId so the auth reducer's seed resets per connection — a new
-          // agent can't inherit the prior session's sign-out gate/in-flight state.
-          key={account ? `account-${account.agentId}` : 'account-none'}
-          account={account}
-          onSignedOut={onSignedOut}
-        />
-      </section>
-      <section className="flex flex-col gap-2">
-        <h2 className="text-[13px] font-semibold text-faint">Environment</h2>
-        <Environment detect={detect} loading={loading} onRecheck={onRecheck} />
-      </section>
-    </div>
-  )
-}
-
-/** The selected Workspace's connected agent + its advertised auth, for the Account section. */
-interface AccountInfo {
-  agentId: string
-  authMethods: AuthMethod[]
-  signOutAvailable: boolean
-}
-
-/**
- * The Settings > Account section (moved off the old chat banner). Shows the signed-in
- * status for the selected Workspace's warm agent + a design-system Sign-out control
- * gated on `signOutAvailable`, mirroring the old `SignedInBar`'s pure `authReducer` /
- * `signOut` lifecycle but styled with tokens + `Button` (no legacy banner BEM). Sign-out
- * still calls `window.api.signOut({ agentId })` and, on success, routes that Workspace
- * to its sign-in panel via `onSignedOut` (which also closes Settings). When no Workspace
- * is connected (`account` null) it shows a muted hint and offers no Sign-out.
- */
-function AccountSettings({
-  account,
-  onSignedOut,
-}: {
-  account: AccountInfo | null
-  onSignedOut: (authMethods: AuthMethod[]) => void
-}): JSX.Element {
-  const [state, dispatch] = useReducer(
-    authReducer,
-    signedInAuthViewState(account?.authMethods ?? [], account?.signOutAvailable ?? false),
-  )
-  const view = selectAuthView(state)
-
-  async function signOut(): Promise<void> {
-    if (!account) return
-    dispatch({ type: 'sign-out-start' })
-    const result = await window.api.signOut({ agentId: account.agentId })
-    if (result.ok) {
-      dispatch({ type: 'sign-out-success' })
-      onSignedOut(result.authMethods)
-    } else {
-      dispatch({ type: 'sign-out-error', message: result.error })
-    }
-  }
-
-  if (!account) {
-    return (
-      <div className="rounded-[9px] border border-border p-3 text-[13px] text-muted">
-        Not connected — open a project to manage your session.
-      </div>
-    )
-  }
-
-  const signingOut = view.kind === 'signing-out'
-  return (
-    <div className="flex flex-col gap-2.5 rounded-[9px] border border-border p-3">
-      <div className="flex items-center gap-2 text-[13px]">
-        {!signingOut && <span className="size-[7px] shrink-0 rounded-full bg-ok" aria-hidden />}
-        <span className="font-semibold text-text-strong">
-          {signingOut ? 'Signing out…' : 'Signed in to Mistral Vibe'}
-        </span>
-        {view.kind === 'signed-in' && view.identity && (
-          <span className="text-muted">{view.identity}</span>
-        )}
-        <span className="flex-1" />
-        {view.kind === 'signed-in' && view.signOutAvailable && (
-          <Button variant="outline" size="sm" onClick={() => void signOut()}>
-            Sign out
-          </Button>
-        )}
-      </div>
-      {view.kind === 'signed-in' && view.error && (
-        <div className="text-[13px] text-bad">{view.error}</div>
-      )}
-    </div>
-  )
-}
-
-/** The environment check: whether `vibe` / `vibe-acp` are installed + reachable. */
-function Environment({
-  detect,
-  loading,
-  onRecheck,
-}: {
-  detect: VibeDetectResult | null
-  loading: boolean
-  onRecheck: () => void
-}): JSX.Element {
-  return (
-    <div className="flex flex-col gap-2.5 rounded-[9px] border border-border p-3">
-      <div className="flex items-center justify-between text-[13px] font-semibold text-text-strong">
-        <span>Environment</span>
-        <Button variant="ghost" size="xs" onClick={onRecheck} disabled={loading}>
-          {loading ? 'Checking…' : 'Re-check'}
-        </Button>
-      </div>
-      {detect && (
-        <ul className="status">
-          <StatusRow ok={detect.vibeFound} label="vibe CLI" />
-          <StatusRow ok={detect.vibeAcpFound} label="vibe-acp (ACP server)" />
-          <li className="status__row">
-            <span className="status__label">version</span>
-            <span className="status__value">{detect.vibeVersion ?? '—'}</span>
-          </li>
-          {detect.error && <li className="status__error">{detect.error}</li>}
-        </ul>
-      )}
-    </div>
-  )
-}
-
-/**
- * The not-signed-in panel: clicking Sign-in drives Vibe's delegated browser
- * sign-in via main; on success it bubbles up (`onSignedIn`) so the app opens a
- * Thread on the same retained agent and lands in a connected conversation. The
- * auth lifecycle (signing-in / signed-in / error) is the pure `authReducer`.
- */
-function SignInPanel({
-  agentId,
-  authMethods,
-  onSignedIn,
-}: {
-  agentId: string
-  authMethods: AuthMethod[]
-  onSignedIn: () => void
-}): JSX.Element {
-  const [state, dispatch] = useReducer(authReducer, authMethods, initialAuthViewState)
-  // Generation counter: bumped on every attempt start and on cancel. Neither the
-  // delegated `complete` long-poll nor the blocking `browser-auth` call can be
-  // aborted over ACP, so a cancelled (or superseded) attempt's eventual result
-  // must be ignored rather than clobber the panel — we only apply a result whose
-  // generation is still current. (This guard is method-agnostic: it covers both
-  // the delegated primary and the blocking fallback.)
-  const attemptRef = useRef(0)
-  const view = selectAuthView(state)
-
-  async function signIn(methodId: string): Promise<void> {
-    const attempt = ++attemptRef.current
-    dispatch({ type: 'sign-in-start' })
-    const result = await window.api.signIn({ agentId, methodId })
-    if (attempt !== attemptRef.current) return // cancelled/superseded — drop the stale result
-    if (result.ok && result.authState === 'signed-in') {
-      dispatch({ type: 'sign-in-success' })
-      onSignedIn() // continue to a connected Thread on the same agent
-    } else {
-      dispatch({
-        type: 'sign-in-error',
-        message: result.ok ? 'Sign-in did not complete. Please try again.' : result.error,
-      })
-    }
-  }
-
-  function cancel(): void {
-    attemptRef.current++ // invalidate the in-flight attempt; its result is dropped
-    dispatch({ type: 'sign-in-cancel' })
-  }
-
-  // #79: OBSERVE current auth state without re-running the browser flow — recovers
-  // an out-of-band `vibe` CLI sign-in, the blocking fallback, or a delegated
-  // `complete` whose result we lost. Bumps the attempt generation so any stale
-  // in-flight sign-in result is dropped (same guard as `signIn`/`cancel`).
-  async function checkStatus(): Promise<void> {
-    const attempt = ++attemptRef.current
-    dispatch({ type: 'check-start' })
-    const result = await window.api.checkAuthStatus({ agentId })
-    if (attempt !== attemptRef.current) return // superseded — drop the stale result
-    if (result.ok && result.authState === 'signed-in') {
-      dispatch({ type: 'sign-in-success' })
-      onSignedIn() // continue to a connected Thread on the same agent
-    } else if (result.ok) {
-      dispatch({
-        type: 'sign-in-error',
-        message: 'Still not signed in. Finish signing in (in your browser or via `vibe`), then check again.',
-      })
-    } else {
-      dispatch({ type: 'sign-in-error', message: result.error })
-    }
-  }
-
-  if (view.kind === 'signed-in') {
-    return (
-      <SignInCard>
-        <SignInLoading label="Signed in — opening your workspace…" />
-      </SignInCard>
-    )
-  }
-
-  if (view.kind === 'signing-out') {
-    return (
-      <SignInCard>
-        <SignInLoading label="Signing out…" />
-      </SignInCard>
-    )
-  }
-
-  if (view.kind === 'signing-in') {
-    return (
-      <SignInCard>
-        <SignInLoading label="Signing in…" />
-        <p className="text-[13px] leading-relaxed text-muted">
-          Complete sign-in in your browser, then return here.
-        </p>
-        <Button variant="outline" size="sm" className="self-start" onClick={cancel}>
-          Cancel
-        </Button>
-      </SignInCard>
-    )
-  }
-
-  if (view.kind === 'checking') {
-    return (
-      <SignInCard>
-        <SignInLoading label="Checking sign-in status…" />
-      </SignInCard>
-    )
-  }
-
-  // sign-in or error: both render the (clickable) Sign-in button so the error
-  // state stays recoverable, plus a re-check that OBSERVES auth state without
-  // re-running the browser flow (#79) — for an out-of-band `vibe` CLI sign-in.
-  return (
-    <SignInCard>
-      <div className="text-sm font-semibold text-text-strong">Not signed in to Mistral Vibe</div>
-      {view.kind === 'sign-in' && view.description && (
-        <p className="text-[13px] leading-relaxed text-muted">{view.description}</p>
-      )}
-      {view.kind === 'error' && <p className="text-[13px] leading-relaxed text-bad">{view.message}</p>}
-      <div className="flex flex-wrap items-center gap-2">
-        <Button variant="default" onClick={() => void signIn(view.methodId)}>
-          {view.kind === 'error' ? `Retry — ${view.methodName}` : view.methodName}
-        </Button>
-        <Button variant="ghost" onClick={() => void checkStatus()}>
-          Already signed in? Check status
-        </Button>
-      </div>
-    </SignInCard>
-  )
-}
-
-/** Centered card chrome shared by every {@link SignInPanel} state (sibling of the
- * connecting/error connect-states). Token surface + rounded card, no legacy BEM. */
-function SignInCard({ children }: { children: ReactNode }): JSX.Element {
-  return <Card className="mx-auto mt-14 w-full max-w-[420px] gap-3">{children}</Card>
-}
-
-/** A branded-spinner + label row for the panel's in-flight states (signing-in /
- * checking / signing-out / opening). Copy is unchanged; only the look. */
-function SignInLoading({ label }: { label: string }): JSX.Element {
-  return (
-    <div className="flex items-center gap-2.5 text-sm font-semibold text-text-strong">
-      {/* Spinner is decorative here — the visible label carries the accessible name,
-          so hide the spinner's own role=img to avoid a double screen-reader announce. */}
-      <span aria-hidden className="inline-flex">
-        <LogoSnakeSpinner size={16} />
-      </span>
-      <span>{label}</span>
-    </div>
-  )
-}
-
 /** The persisted Threads under a connected Workspace (by minted id), for its list (TB5). */
 function threadsForWorkspace(recents: ListMetadataResult, workspaceId: string): ThreadMeta[] {
   return recents.find((w) => w.id === workspaceId)?.threads ?? []
-}
-
-function StatusRow({ ok, label }: { ok: boolean; label: string }): JSX.Element {
-  return (
-    <li className="status__row">
-      <span className={ok ? 'dot dot--ok' : 'dot dot--bad'} aria-hidden />
-      <span className="status__label">{label}</span>
-      <span className="status__value">{ok ? 'found' : 'missing'}</span>
-    </li>
-  )
 }
